% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{accuracy}
\alias{accuracy}
\title{Compute accuracy}
\usage{
accuracy(pred, truth)
}
\arguments{
\item{pred}{A logical vector of predictions.}

\item{truth}{A logical vector of true values.}
}
\value{
A numeric value (between 0 and 1) representing the accuracy of the predictions.
}
\description{
Compute accuracy
}
\examples{
accuracy(c(TRUE, TRUE, FALSE, FALSE), c(TRUE, FALSE, TRUE, FALSE))
}
