% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{adj_R2}
\alias{adj_R2}
\title{Compute adjusted R-squared}
\usage{
adj_R2(pred, truth, n_p)
}
\arguments{
\item{pred}{A numeric vector of predictions.}

\item{truth}{A numeric vector of true values.}

\item{n_p}{The number of predictors.}
}
\value{
A numeric value (between 0 and 1) representing the adjusted R-squared of the predictions against the true values.
}
\description{
Compute adjusted R-squared
}
\examples{
adj_R2(c(1, 2, 3, 4, 5), c(1, 2, 3, 4, 5), n_p = 1)
}
