% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMIPackage.R
\docType{package}
\name{BMIPackage}
\alias{BMIPackage}
\title{Biostatistics for Machine Learning Toolbox}
\description{
A toolbox for biostatistics and machine learning functions. The R package provides a collection of 220 convenient utility functions for data analysis and machine learning tasks.  It offers a variety of random sampling, statistical metrics calculation, and data manipulation functions. Several functions compute log-likelihoods for various statistical distributions, and others calculate performance metrics such as accuracy, recall, and precision.  Overall, this package aims to streamline and simplify common tasks in data analysis and predictive modeling.
}
\author{
Elizabeth Nemeti \email{enemeti@emory.edu}
}
