% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{classes}
\alias{classes}
\title{Extract classes of columns}
\usage{
classes(x)
}
\arguments{
\item{x}{A data frame.}
}
\value{
A character vector of class names for each column in x.
}
\description{
Extract classes of columns
}
\examples{
classes(data.frame(a = 1:3, b = letters[1:3]))
}
