% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{df_scale}
\alias{df_scale}
\title{Scale a data frame}
\usage{
df_scale(x, center = T, scale = T)
}
\arguments{
\item{x}{A data frame.}

\item{center}{Logical, whether to center the columns.}

\item{scale}{Logical, whether to scale the columns.}
}
\value{
A data frame where the numeric columns have been scaled (centered and/or standardized).
}
\description{
Scale a data frame
}
\examples{
df_scale(data.frame(a = 1:3, b = 4:6))
}
