% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{f1}
\alias{f1}
\title{Compute F1 score}
\usage{
f1(pred, truth)
}
\arguments{
\item{pred}{A logical vector of predictions.}

\item{truth}{A logical vector of true values.}
}
\value{
A numeric value (between 0 and 1) representing the F1 score of the predictions.
}
\description{
Compute F1 score
}
\examples{
f1(c(TRUE, TRUE, FALSE, FALSE), c(TRUE, FALSE, TRUE, FALSE))
}
