% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{is_max}
\alias{is_max}
\title{Identify maximum value}
\usage{
is_max(x, na.rm = T)
}
\arguments{
\item{x}{A numeric vector.}

\item{na.rm}{Logical, whether to remove NA values.}
}
\value{
A logical vector indicating the maximum values in x (TRUE if the value is a maximum, FALSE otherwise).
}
\description{
Identify maximum value
}
\examples{
is_max(c(1, 2, 3, NA), na.rm = TRUE)
}
