% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{is_min}
\alias{is_min}
\title{Identify minimum value}
\usage{
is_min(x, na.rm = T)
}
\arguments{
\item{x}{A numeric vector.}

\item{na.rm}{Logical, whether to remove NA values.}
}
\value{
A logical vector indicating the minimum values in x (TRUE if the value is a minimum, FALSE otherwise).
}
\description{
Identify minimum value
}
\examples{
is_min(c(1, 2, 3, NA), na.rm = TRUE)
}
