% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{log_likelihood_chisq}
\alias{log_likelihood_chisq}
\title{Compute log-likelihood for chi-squared distribution}
\usage{
log_likelihood_chisq(x, df)
}
\arguments{
\item{x}{A numeric vector.}

\item{df}{Degrees of freedom.}
}
\value{
A numeric value representing the log-likelihood of the data given the specified chi-squared distribution.
}
\description{
Compute log-likelihood for chi-squared distribution
}
\examples{
log_likelihood_chisq(rchisq(100, df = 1), df = 1)
}
