% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{log_likelihood_f}
\alias{log_likelihood_f}
\title{Compute log-likelihood for F distribution}
\usage{
log_likelihood_f(x, df1, df2)
}
\arguments{
\item{x}{A numeric vector.}

\item{df1}{Numerator degrees of freedom.}

\item{df2}{Denominator degrees of freedom.}
}
\value{
A numeric value representing the log-likelihood of the data given the specified F distribution.
}
\description{
Compute log-likelihood for F distribution
}
\examples{
log_likelihood_f(rf(100, df1 = 1, df2 = 1), df1 = 1, df2 = 1)
}
