% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{log_likelihood_norm}
\alias{log_likelihood_norm}
\title{Compute log-likelihood for normal distribution}
\usage{
log_likelihood_norm(x, mean, sd)
}
\arguments{
\item{x}{A numeric vector.}

\item{mean}{The mean of the normal distribution.}

\item{sd}{The standard deviation of the normal distribution.}
}
\value{
A numeric value representing the log-likelihood of the data given the specified normal distribution.
}
\description{
Compute log-likelihood for normal distribution
}
\examples{
log_likelihood_norm(rnorm(100), mean = 0, sd = 1)
}
