% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{log_likelihood_t}
\alias{log_likelihood_t}
\title{Compute log-likelihood for t distribution}
\usage{
log_likelihood_t(x, df)
}
\arguments{
\item{x}{A numeric vector.}

\item{df}{Degrees of freedom.}
}
\value{
A numeric value representing the log-likelihood of the data given the specified t distribution.
}
\description{
Compute log-likelihood for t distribution
}
\examples{
log_likelihood_t(rt(100, df = 1), df = 1)
}
