% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{log_likelihood_unif}
\alias{log_likelihood_unif}
\title{#' Compute log-likelihood for uniform distribution}
\usage{
log_likelihood_unif(x, min, max)
}
\arguments{
\item{x}{A numeric vector.}

\item{min}{The minimum value of the uniform distribution.}

\item{max}{The maximum value of the uniform distribution.}
}
\value{
A numeric value representing the log-likelihood of the data given the specified uniform distribution.
}
\description{
#' Compute log-likelihood for uniform distribution
}
\examples{
log_likelihood_unif(runif(100), min = 0, max = 1)
}
