% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{minimum_n_per_group}
\alias{minimum_n_per_group}
\title{Compute minimum sample size per group}
\usage{
minimum_n_per_group(d, power = 0.8)
}
\arguments{
\item{d}{The effect size.}

\item{power}{The statistical power.}
}
\value{
An integer value representing the minimum sample size per group needed to detect the effect size with the specified power.
}
\description{
Compute minimum sample size per group
}
\examples{
minimum_n_per_group(d = 0.5, power = 0.8)
}
