% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{precision}
\alias{precision}
\title{Compute precision}
\usage{
precision(pred, truth)
}
\arguments{
\item{pred}{A logical vector of predictions.}

\item{truth}{A logical vector of true values.}
}
\value{
A numeric value (between 0 and 1) representing the precision of the predictions.
}
\description{
Compute precision
}
\examples{
precision(c(TRUE, TRUE, FALSE, FALSE), c(TRUE, FALSE, TRUE, FALSE))
}
