% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{r2}
\alias{r2}
\title{Compute R-squared}
\usage{
r2(pred, truth)
}
\arguments{
\item{pred}{A numeric vector of predictions.}

\item{truth}{A numeric vector of true values.}
}
\value{
A numeric value (between 0 and 1) representing the R-squared of the predictions against the true values.
}
\description{
Compute R-squared
}
\examples{
r2(c(1, 2, 3, 4, 5), c(1, 2, 3, 4, 5))
}
