% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{rando}
\alias{rando}
\title{Select random samples}
\usage{
rando(x, n = 1, replace = T)
}
\arguments{
\item{x}{A vector or a data frame.}

\item{n}{The number of items to select.}

\item{replace}{Logical, whether to sample with replacement.}
}
\value{
A vector or a data frame containing random samples from x, depending on the input type.
}
\description{
Select random samples
}
\examples{
rando(1:10, n = 5)
}
