% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{rep_mat}
\alias{rep_mat}
\title{Repeat matrix}
\usage{
rep_mat(x, M, N)
}
\arguments{
\item{x}{A matrix.}

\item{M}{The number of times to repeat rows.}

\item{N}{The number of times to repeat columns.}
}
\value{
A matrix where the rows and columns of x are repeated M and N times, respectively.
}
\description{
Repeat matrix
}
\examples{
rep_mat(matrix(1:4, nrow = 2), M = 2, N = 3)
}
